requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-genai
python-dotenv
PyJWT
python-multipart
pwdlib